+++ cmd/validate.go
@@ -14,26 +14,24 @@ func NewValidateCmd(configFile *string, reportFile *string) *cobra.Command {
 		Use:   "validate",
 		Short: "compare thresholds in config to findings in report",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			var GateCheckConfig *gatecheck.Config
+			var config *gatecheck.Config
 			var err error
 
 			if flagIgnoreConfig == false {
-				GateCheckConfig, err = OpenAndDecode[gatecheck.Config](*configFile, YAML)
+				config, err = OpenAndDecode[gatecheck.Config](*configFile, YAML)
 				if err != nil {
 					return err
 				}
 			}
 
-			GateCheckReport, err := OpenAndDecode[gatecheck.Report](*reportFile, JSON)
+			report, err := OpenAndDecode[gatecheck.Report](*reportFile, JSON)
 			if err != nil {
 				return err
 			}
 
-			GateCheckReport = GateCheckReport.WithConfig(GateCheckConfig)
+			report = report.WithConfig(config)
 
-			err = gatecheck.NewStdValidator(*GateCheckReport).Validate()
-
-			if err != nil {
+			if err := report.Validate(); err != nil {
 				cmd.PrintErrln(err.Error())
 				if flagAudit == true {
 					return nil