+++ pkg/gatecheck/report.go
@@ -1,11 +1,15 @@
 package gatecheck
 
 import (
+	"errors"
+	"fmt"
 	"github.com/gatecheckdev/gatecheck/pkg/artifact/grype"
 	"github.com/gatecheckdev/gatecheck/pkg/artifact/semgrep"
 	"strings"
 	"time"
 )
+
+var ErrorValidation = errors.New("report failed validation")
 
 type Settings struct {
 	ProjectName string
@@ -53,6 +57,31 @@ func (r Report) WithSettings(s Settings) *Report {
 	return &r
 }
 
+// Validate calls the validate function for each artifact
+func (r Report) Validate() error {
+	var allErrors []error
+	var errorDescriptions []string
+	for _, artifact := range r.artifacts() {
+		if err := artifact.Validate(); err != nil {
+			allErrors = append(allErrors, err)
+			errorDescriptions = append(errorDescriptions, err.Error())
+		}
+	}
+	if len(allErrors) != 0 {
+		return fmt.Errorf("%w : %s", ErrorValidation,
+			strings.Join(errorDescriptions, "\n"))
+	}
+
+	return nil
+}
+
+func (r Report) artifacts() []Artifact {
+	return []Artifact{
+		r.Artifacts.Grype,
+		r.Artifacts.Semgrep,
+	}
+}
+
 func (r Report) String() string {
 	var out strings.Builder
 	divider := strings.Repeat("-", 25) + "\n"
@@ -63,4 +92,9 @@ func (r Report) String() string {
 	out.WriteString(divider)
 	out.WriteString(r.Artifacts.Semgrep.String())
 	return out.String()
+}
+
+type Artifact interface {
+	Validate() error
+	fmt.Stringer
 }