+++ pkg/gatecheck/report_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/gatecheckdev/gatecheck/pkg/artifact/grype"
+	"github.com/gatecheckdev/gatecheck/pkg/artifact/semgrep"
 	"os"
 	"strings"
 	"testing"
@@ -85,4 +86,57 @@ func TestReport_WithSettings(t *testing.T) {
 	if strings.Compare(r.PipelineUrl, "pipeline.com") != 0 {
 		t.Fatal("Unexpected Pipeline URL")
 	}
+}
+
+func TestReport_Validate(t *testing.T) {
+	r := NewReport("Test Report")
+	r.Artifacts.Grype.Critical.Found = 20
+	r.Artifacts.Grype.High.Found = 22
+	r.Artifacts.Grype.Medium.Found = 113
+
+	r.Artifacts.Semgrep.Error.Found = 12
+	r.Artifacts.Semgrep.Warning.Found = 14
+	r.Artifacts.Semgrep.Info.Found = 130
+
+	t.Run("All Vulnerabilities allowed", func(t *testing.T) {
+		// All vulnerabilities should be allowed
+		c := NewConfig("Test Project")
+
+		r = r.WithConfig(c)
+		if err := r.Validate(); err != nil {
+			t.Fatal("Validation should pass")
+		}
+	})
+
+	t.Run("Some Allowed", func(t *testing.T) {
+		c := NewConfig("Test Project")
+		c.Grype.Critical = 0
+		c.Grype.High = 1
+		c.Grype.Medium = 2
+		c.Grype.Low = 50
+		c.Grype.Unknown = 0
+		c.Grype.Negligible = 0
+
+		c.Semgrep.Error = 0
+		c.Semgrep.Warning = 20
+		c.Semgrep.Info = -1
+		r = r.WithConfig(c)
+
+		if err := r.Validate(); err == nil {
+			t.Fatal("Validation should fail")
+		}
+
+	})
+
+	t.Run("No vulnerabilities allowed", func(t *testing.T) {
+		c := NewConfig("Test Project")
+		c.Grype = *grype.NewConfig(0)
+		c.Semgrep = *semgrep.NewConfig(0)
+		r = r.WithConfig(c)
+
+		if err := r.Validate(); err == nil {
+			t.Fatal("Validation should fail")
+		}
+	})
+
 }