+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor unit tests in cmd package to be more uniform
 - Removed test utility and internal packages in favor of IO functions
 - Move config, report, and validator to pkg/gatecheck for simplified folder structure
+- Moved validate responsibility to the artifact
 
 ### Added
 - JSON struct tags to config for additional support